functions/save-qr/requirements.txt
functions-framework==3.*
qrcode[pil]==7.*
google-cloud-storage==2.*
